move Global to CustomBehaviors
