+++ SolastaUnfinishedBusiness/Api/GameExtensions/RulesetSpellRepertoireExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using SolastaUnfinishedBusiness.Api.Helpers;
+using SolastaUnfinishedBusiness.CustomBehaviors;
 using SolastaUnfinishedBusiness.Models;
 
 namespace SolastaUnfinishedBusiness.Api.GameExtensions;
