+++ SolastaUnfinishedBusiness/Displays/EncountersDisplay.cs
@@ -4,6 +4,7 @@ using System.Text.RegularExpressions;
 using JetBrains.Annotations;
 using SolastaUnfinishedBusiness.Api.LanguageExtensions;
 using SolastaUnfinishedBusiness.Api.ModKit;
+using SolastaUnfinishedBusiness.CustomBehaviors;
 using SolastaUnfinishedBusiness.Models;
 
 namespace SolastaUnfinishedBusiness.Displays;
