+++ SolastaUnfinishedBusiness/Displays/GameUiDisplay.cs
@@ -1,4 +1,5 @@
 ﻿using SolastaUnfinishedBusiness.Api.ModKit;
+using SolastaUnfinishedBusiness.CustomBehaviors;
 using SolastaUnfinishedBusiness.CustomUI;
 using SolastaUnfinishedBusiness.Models;
 using UnityEngine;
