+++ SolastaUnfinishedBusiness/Models/CharacterExportContext.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using JetBrains.Annotations;
+using SolastaUnfinishedBusiness.CustomBehaviors;
 using TMPro;
 using UnityEngine;
 using Object = UnityEngine.Object;
