+++ SolastaUnfinishedBusiness/Models/InventoryManagementContext.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using JetBrains.Annotations;
+using SolastaUnfinishedBusiness.CustomBehaviors;
 using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
