+++ SolastaUnfinishedBusiness/Models/InvocationsContext.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using JetBrains.Annotations;
+using SolastaUnfinishedBusiness.CustomBehaviors;
 using UnityEngine;
 using UnityEngine.UI;
 using static SolastaUnfinishedBusiness.CustomBuilders.InvocationsBuilders;
