+++ SolastaUnfinishedBusiness/Models/ToolsContext.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using JetBrains.Annotations;
 using SolastaUnfinishedBusiness.Api.Helpers;
 using SolastaUnfinishedBusiness.Builders;
+using SolastaUnfinishedBusiness.CustomBehaviors;
 using TA;
 using UnityEngine;
 using UnityEngine.UI;
